ENH Begin move to generic DataRequest system - allow fancy literal overload for specifying format-specific requests and positional arguments. Begin work to specify stream variants vai the locators. Fix some minor warnings.

+++ tests/cxx/test_random_patterns.cc
@@ -109,7 +109,7 @@ TEST_F(RandomPatternTest, FixedPatternFill) {
                                    ncarray::DType::uint8,
                                    0,
                                    false);
-    std::size_t total_bytes { bytes_arr.nbytes() };
+    std::size_t total_bytes { static_cast<std::size_t>(bytes_arr.nbytes()) };
 
     for (std::size_t i = 0; i < total_bytes; ++i) {
       std::uint8_t& val { bytes_arr[{0, i}] };
@@ -150,7 +150,7 @@ TEST_F(RandomPatternTest, SequentialPatternFill) {
                                    ncarray::DType::uint8,
                                    0,
                                    false);
-    std::size_t total_bytes { bytes_arr.nbytes() };
+    std::size_t total_bytes { static_cast<std::size_t>(bytes_arr.nbytes()) };
 
     for (std::size_t i = 0; i < total_bytes; ++i) {
       std::uint8_t expected { static_cast<std::uint8_t>((step + i) & 0xFF) };
@@ -194,7 +194,7 @@ TEST_F(RandomPatternTest, PRNGPatternFillDeterminism) {
                                    ncarray::DType::uint8,
                                    0,
                                    false);
-    std::size_t total_bytes { bytes_arr.nbytes() };
+    std::size_t total_bytes { static_cast<std::size_t>(bytes_arr.nbytes()) };
 
     // Verify LCG output formula: state = state * 1664525 + 1013904223
     std::uint32_t state { seed ^ static_cast<std::uint32_t>(step * 0x9E3779B9u + 1) };
